fix wrong type of memcached_instance_st::server_timeout_counter_query_id

+++ libmemcached/instance.hpp
@@ -149,7 +149,7 @@ struct memcached_instance_st {
   uint32_t server_failure_counter;
   uint64_t server_failure_counter_query_id;
   uint32_t server_timeout_counter;
-  uint32_t server_timeout_counter_query_id;
+  uint64_t server_timeout_counter_query_id;
   uint32_t weight;
   uint32_t version;
   enum memcached_server_state_t state;
